4.13

+++ msgHandler/RoleQuitRoomHandler.go
@@ -15,22 +15,28 @@ func (handler *RoleQuitRoomHandler) Action(session network.SocketSessionInterfac
 	if innerMsg, ok := msg.(network.InnerWriteMessage); ok {
 		if _, ok := innerMsg.MsgData.(*message.C2M_QuitRoom); ok {
 			role := handler.GameServer.GetRoleManager().GetOnlineRole(innerMsg.RoleId)
+			if role.IsInBattling(){
+				//如果在战斗中就不退出
+				return
+			}
 			roomManager := handler.GameServer.GetRoomManager()
-			room := roomManager.GetRoomByRoomId(role.GetRoomId())
-			roomId := room.GetRoomId()
+			roomId := role.GetRoomId()
+			room := roomManager.GetRoomByRoomId(roomId)
 			if room != nil{
 				//如果是房主退出就地解散
 				if room.IsRoomOwner(role.GetRoleId()){
 					if roomManager.DeleteRoom(roomId){
 						log4g.Infof("删除房间[%d]成功!",roomId)
 					}else{
-						log4g.Errorf("删除房间[%d]失败!",roomId)
+						log4g.Infof("删除房间[%d]失败!",roomId)
 					}
 				}else{
 					if roomManager.RemoveOneMemberByRoom(room,role) == false{
-						log4g.Errorf("移除房间[%d]成员[%d]失败!",roomId,role.GetRoleId())
+						log4g.Infof("移除房间[%d]成员[%d]失败!",roomId,role.GetRoleId())
 					}
 				}
+			}else {
+				log4g.Infof("房间[%d]已经不存在了!",roomId)
 			}
 		}
 	}
